cleanup package hierarchy
1. add "music.canon" with DSL, StdLib, and "canonical" string
   representations (formerly included in Music.kt).
2. add READMEs for all packages.
