+++ src/main/kotlin/org/pareto/music/canon/README.md
@@ -1,0 +1,3 @@
+# `music.canon`
+
+This package contains the standard library and tools for composing `music`.
