+++ src/main/kotlin/org/pareto/music/demo/README.md
@@ -1,0 +1,3 @@
+# `music.demo`
+
+This package contains example use cases for `Music`. These subpackages should be standalone and should not be internally imported. Think of them as lightweight little product demos and please add more!
