+++ src/main/kotlin/org/pareto/music/rehearse/Compiler.kt
@@ -7,6 +7,7 @@ import org.pareto.music.Decision
 import org.pareto.music.Melody
 import org.pareto.music.Harmony
 import org.pareto.music.Note
+import org.pareto.music.Silence
 
 
 typealias Deferred<T> = () -> T
